CLB does not use unique names
Support the fact that Cloud Load Balancers do not require unique names
to identify them. We'll use this fact to create multiple load balancers
with the same IP address but on different ports with the same name. To
allow for this understand that we can get multiple load balancers back
with the same name and find the right one by port.