support multiple ports per service
Cloud Load Balancers do not have a listener port definition but instead
we must create a load balancer per port that we want to listen on.
Adjust the code to loop through each port on the service and create a
load balancer for each one. Ensure when a port on the service is removed
that we remove the correct load balancer.